Add `eager_mode()` definition and call in `__init__()`
New work on BB2 will create a `LazyJLLWrappers` that has a (slightly)
different API, and is lazily-loaded.  In order to be compatible with
JLLs that still use `JLLWrappers`, we define an `eager_mode()` method
that eagerly loads all libraries for use in a downstream JLL that still
uses `JLLWrappers`.

This PR adds no-op `eager_mode()` implementations for all old JLLs, and
invokes `eager_mode()` on all dependencies to pave the way for this
interopability.