Avoid side effects when precompiling `find_artifact_dir`
With an explicit call to `find_artifact_dir()` we were actually fetching lazy
artifacts during precompilation, which is typically undesired, besides hitting a
bug in Downloads.jl in certain versions of Julia (up to 1.10.4) which causes the
download of the artifact to hang.  Instead, with `Base.precompile` we retain the
caching of native code, without causing extra downloads during precompilation.